Added some images. Sum showing bug fixed.

+++ MoneyIO/app/src/main/java/io/money/moneyio/model/recyclers/HistoryRecyclerViewAdapter.java
@@ -59,7 +59,7 @@ public class HistoryRecyclerViewAdapter extends RecyclerView.Adapter<HistoryRecy
 
         holder.comment.setText(moneyFlow.getComment());
         holder.type.setText(moneyFlow.getType());
-        holder.price.setText(Double.toString(moneyFlow.getSum()));
+        holder.price.setText(Float.toString(moneyFlow.getSum()));
         holder.date.setText(new SimpleDateFormat("d-MMM-yy' / 'HH:mm").format(new Date(moneyFlow.getCalendar())));
         if (moneyFlow.getExpense().equals("true")) {
             holder.image.setImageResource(R.drawable.outcome_icon);//outcome
